# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Buyer Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/user",
    "paths": {
        "/customer/login": {
            "post": {
                "description": "Return JWT access and refresh pair, alongside user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "BuyerLogin user",
                "parameters": [
                    {
                        "description": "BuyerLogin input",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UserSignIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.LoginResponse"
                        }
                    }
                }
            }
        },
        "/customer/profile": {
            "get": {
                "security": [
                    {
                        "JWT Key": []
                    }
                ],
                "description": "Get customer profile from Authorization JWT header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Get Buyer BuyerProfile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer YourJWTToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.UserResponse"
                        }
                    }
                }
            }
        },
        "/customer/refresh_token": {
            "post": {
                "description": "Return JWT access token given refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Refresh token handler",
                "parameters": [
                    {
                        "description": "Receive refresh token",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customer/register": {
            "post": {
                "description": "Register buyer account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Register customer",
                "parameters": [
                    {
                        "description": "Signup input",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UserSignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.LoginResponse"
                        }
                    }
                }
            }
        },
        "/seller/login": {
            "post": {
                "description": "Return JWT access and refresh pair, alongside user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "SellerLogin user",
                "parameters": [
                    {
                        "description": "SellerLogin input",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UserSignIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.LoginResponse"
                        }
                    }
                }
            }
        },
        "/seller/profile": {
            "get": {
                "security": [
                    {
                        "JWT Key": []
                    }
                ],
                "description": "Get seller profile from Authorization JWT header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Get Seller SellerProfile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer YourJWTToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.UserResponse"
                        }
                    }
                }
            }
        },
        "/seller/refresh_token": {
            "post": {
                "description": "Return JWT access token given refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Refresh token handler",
                "parameters": [
                    {
                        "description": "Receive refresh token",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/seller/register": {
            "post": {
                "description": "Register seller account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Register customer",
                "parameters": [
                    {
                        "description": "Signup input",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UserSignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.LoginResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "forms.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/forms.UserResponse"
                }
            }
        },
        "forms.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "forms.UserGroupResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "forms.UserProfileResponse": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "forms.UserResponse": {
            "type": "object",
            "properties": {
                "group": {
                    "$ref": "#/definitions/forms.UserGroupResponse"
                },
                "id": {
                    "type": "integer"
                },
                "profile": {
                    "$ref": "#/definitions/forms.UserProfileResponse"
                },
                "username": {
                    "type": "string"
                },
                "wallet_balance": {
                    "type": "integer"
                }
            }
        },
        "forms.UserSignIn": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "forms.UserSignUp": {
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "password",
                "username"
            ],
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth  Authorization": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}